+++ src/layer/TileLayer.js
@@ -35,12 +35,7 @@ L.TileLayer = L.Class.extend({
 		this._initContainer();
 		
 		// create an image to clone for tiles
-		this._tileImg = L.DomUtil.create('img', 'leaflet-tile');
-		this._tileImg.galleryimg = 'no';
-		
-		var tileSize = this.options.tileSize;
-		this._tileImg.style.width = tileSize + 'px';
-		this._tileImg.style.height = tileSize + 'px';
+		this._createTileProto();
 		
 		// set up events
 		map.on('viewreset', this._reset, this);
@@ -102,10 +97,10 @@ L.TileLayer = L.Class.extend({
 				Math.floor(bounds.max.y / tileSize)),
 			tileBounds = new L.Bounds(nwTilePoint, seTilePoint);
 		
-		this._loadTilesFromCenterOut(tileBounds);
+		this._addTilesFromCenterOut(tileBounds);
 		
 		if (this.options.unloadInvisibleTiles) {
-			this._unloadOtherTiles(tileBounds);
+			this._removeOtherTiles(tileBounds);
 		}
 	},
 	
@@ -120,7 +115,7 @@ L.TileLayer = L.Class.extend({
 				.replace('{y}', tilePoint.y);
 	},
 	
-	_loadTilesFromCenterOut: function(bounds) {
+	_addTilesFromCenterOut: function(bounds) {
 		var queue = [],
 			center = bounds.getCenter();
 		
@@ -138,11 +133,11 @@ L.TileLayer = L.Class.extend({
 		
 		this._tilesToLoad = queue.length;
 		for (var k = 0, len = this._tilesToLoad; k < len; k++) {
-			this._loadTile(queue[k]);
+			this._addTile(queue[k]);
 		}
 	},
 	
-	_unloadOtherTiles: function(bounds) {
+	_removeOtherTiles: function(bounds) {
 		var kArr, x, y, key;
 		
 		for (key in this._tiles) {
@@ -162,10 +157,8 @@ L.TileLayer = L.Class.extend({
 		}		
 	},
 	
-	_loadTile: function(tilePoint) {
-		var origin = this._map.getPixelOrigin(),
-			tileSize = this.options.tileSize,
-			tilePos = tilePoint.multiplyBy(tileSize).subtract(origin),
+	_addTile: function(tilePoint) {
+		var tilePos = this._getTilePos(tilePoint);
 			zoom = this._map.getZoom();
 			
 		// wrap tile coordinates
@@ -174,42 +167,68 @@ L.TileLayer = L.Class.extend({
 		if (tilePoint.y < 0 || tilePoint.y >= tileLimit) { return; }
 		
 		// create tile
-		var tile = this._tileImg.cloneNode(false);
-		
+		var tile = this._createTile();
 		L.DomUtil.setPosition(tile, tilePos);
 		
 		this._tiles[tilePoint.x + ':' + tilePoint.y] = tile;
 
-		tile._leaflet_layer = this;
-		tile.onload = this._tileOnLoad;
-		tile.onerror = this._tileOnError;
-		tile.onselectstart = tile.onmousemove = L.Util.falseFn;
-		
-		tile.src = this.getTileUrl(tilePoint, zoom);
+		this._loadTile(tile, tilePoint, zoom);
 		
 		this._container.appendChild(tile);
 	},
 	
-	_tileOnLoad: function() {
-		this.className += ' leaflet-tile-loaded'; //TODO DomEvent#addListener target 
+	_getTilePos: function(tilePoint) {
+		var origin = this._map.getPixelOrigin(),
+			tileSize = this.options.tileSize;
+		
+		return tilePoint.multiplyBy(tileSize).subtract(origin);
+	},
+	
+	// image-specific code (override to implement e.g. Canvas or SVG tile layer)
+	
+	_createTileProto: function() {
+		this._tileImg = L.DomUtil.create('img', 'leaflet-tile');
+		this._tileImg.galleryimg = 'no';
+		
+		var tileSize = this.options.tileSize;
+		this._tileImg.style.width = tileSize + 'px';
+		this._tileImg.style.height = tileSize + 'px';
+	},
+	
+	_createTile: function() {
+		var tile = this._tileImg.cloneNode(false);
+		tile._leaflet_layer = this;
+		tile.onselectstart = tile.onmousemove = L.Util.falseFn;
+		return tile;
+	},
+	
+	_loadTile: function(tile, tilePoint, zoom) {
+		tile.onload = L.Util.bind(this._tileOnLoad, this);
+		tile.onerror = L.Util.bind(this._tileOnError, this);
+		tile.src = this.getTileUrl(tilePoint, zoom);
+	},
+	
+	_tileOnLoad: function(e) {
+		var tile = L.DomEvent.getTarget(e);
+		
+		tile.className += ' leaflet-tile-loaded'; 
 
-		var layer = this._leaflet_layer;
+		this.fire('tileload', {tile: tile, url: tile.src});
 		
-		layer.fire('tileload', {tile: this, url: this.src});
-		
-		layer._tilesToLoad--;
-		if (!layer._tilesToLoad) {
-			layer.fire('load');
+		this._tilesToLoad--;
+		if (!this._tilesToLoad) {
+			this.fire('load');
 		}
 	},
 	
-	_tileOnError: function() {
-		this._leaflet_layer.fire('tileerror', {tile: this, url: this.src});
+	_tileOnError: function(e) {
+		var tile = L.DomEvent.getTarget(e);
 		
-		var newUrl = this._leaflet_layer.options.errorTileUrl;
+		this.fire('tileerror', {tile: tile, url: tile.src});
+		
+		var newUrl = this.options.errorTileUrl;
 		if (newUrl) {
-			this.src = newUrl;
+			tile.src = newUrl;
 		}
-			
 	}
 });